Add user selected libdir to the runtime search path
Given autotools flexibility, libraries can be installed in several
locations, even non-standard ones. If the latter happens, GstD will
not be able to find the libraries. By adding LIBDIR to the runtime
search path of the binary we ensure that, no matter where the
libraries are installed, they will be found by GstD.