Adding pointer guards to GstdIpc methods
Summary:
Since session is a member of the GstdIpc struct and should be use in every IPC, we should save a reference of the session here, rather than having the subclass save it for ourselves.

Remove the ipc_ prefix from the methods of the class It's understood that the method is for the class ipc , so no need for the extra prefix.

Forward the error code to the application

Differential Revision: http://phabricator.cr.ridgerun.com/D38