Mark a read property in a GstdProperty
If the read resource is not a GstdObject build a GstdProperty on top of
it so that clients can access it generically. Fundamental types are
directly masked. For derived types, however, we need to get the object
and ask if it is GstdObject based. If it is, then we can just return the
resource. Otherwise, it can be masked as well.