Fix stack overflow corruption
When moving to the threaded socket service an occational seg fault
crashed the server. This segfault was determined to be caused by a stack
overflow where suficiently deep function call chains ran out of stack
memory, corrupting data.

The buffer to hold the TCP message was allocated from stack memory, and
was allocated big enough to hold both request and answer. This memory
was moved to the heap. Additionally, instead of sending back the full
buffer, only N bytes where sent (where N is the length of the response).