Only allow strictly correct pipeline creation
The launch-syntax compiler is capable of creating half-built pipelines
if the description happens to contain errors. This can be problematic if
we later assume that the pipeline was successfully built and is ready to
play. This change adds a flag to make the parsing strict and fail upon
any error.