Ensure the session singleton is NULLed out after disposal
If the session singleton pointer isn't set to NULL after the last
reference is decremented, the next constructor invocation won't create
a new session, but will try to get a reference to an old session which
only containse remaining garbage.

By adding the address of the singleton as a weak reference to the same
singleton, it is guaranteed that upon disposal, that weak pointer will
be set to NULL. This forces the next constructor to generate a session
all over.