Encapsulate pipeline state in a separate object
Previously states were handled as a regular Enum property. This worked
fine except for the fact that property updates have to way to return an
error code. As such, the state was encapsulated in an independent
GstdState implementing its own updater which now forwards the result of
the pipeline state change process.