Create the areas for the pid and log files
PID and log files are typically installed under /var/run and
/var/log respectively. The problem is that, by default, these
areas are write-protected to the regular user. This would
require the user to start gstd using sudo, which is undesired.
In order to avoid this, a new of areas for the PID and log files
are created under /var/run/gstd and /var/log/gstd respectively
and granted write permissions.