Use a less invasive malloc mocking
In newer implementations malloc is declared as a weak symbol and, as such, may be
overriden with a custom implementation. In some of these, as glibc, one needs
to provide a full implementation (malloc, calloc, realloc and free) or undefined
runtime behaviour may raise.

In order to be less invasive, the good old pre-processor definition trick is
used. malloc is substituted to mock_malloc, and the stock malloc (and friends)
may remain the same.