Capture bus message in order to signal a timeout
It is very useful for clients to know if a message was actually
received in the bus, or if it hit a timeout. For this reason we
parse the received message and if the response field is NULL then
we can safely assume that a timeout was hit, and return the
respective error code.